backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   outcome
    #   trio
bcrypt==4.3.0
    # via passlib
black==25.1.0
    # via -r requirements.in
certifi==2025.4.26
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   uvicorn
cryptography==45.0.4
    # via python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.109.2
    # via -r requirements.in
flake8==7.2.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   uvicorn
    #   wsproto
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   trio
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   black
    #   pytest
    #   webdriver-manager
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.13.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.6.1
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.2
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via pytest
pysocks==1.7.1
    # via urllib3
pytest==8.4.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
    #   webdriver-manager
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
requests==2.31.0
    # via
    #   -r requirements.in
    #   webdriver-manager
rsa==4.9.1
    # via python-jose
selenium==4.18.1
    # via -r requirements.in
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.27
    # via
    #   -r requirements.in
    #   alembic
starlette==0.36.3
    # via fastapi
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.14.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   selenium
    #   sqlalchemy
urllib3[socks]==2.4.0
    # via
    #   requests
    #   selenium
uvicorn==0.27.1
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
webdriver-manager==4.0.1
    # via -r requirements.in
websockets==15.0.1
    # via uvicorn
wsproto==1.2.0
    # via trio-websocket
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
tenacity==8.2.3
# LLM dependencies for hybrid book parser
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For efficient model loading
sentencepiece>=0.1.99  # Required for Mistral tokenizer
